requirements.txt
gevent
gevent-websocket
six
ujson
orjson
psycopg2-binary
arrow
openpyxl
wsaccel
whitenoise
beaker
schedule
urllib3
certifi
cryptography
xmltodict
segno
qrcode
pillow
lxml
xlsxwriter
requests
bcrypt
firebase_admin
boto3
brotli
python-docx
phonenumbers
address-hammer
usaddress
zeep
cffi
paramiko
pycparser
pynacl
pysftp
pytz
telnyx
plaid-python
google-pay-token-decryption
applepay
dictdiffer
firebase-admin
langchain
langchain-community
ollama
geopy
opencv-python
opencv-contrib-python
zxing-cpp
pyzbar
pyopenssl
pytesseract
snowflake-connector-python
fastapi
uvicorn
llama-hub
llama-index
llama-index-agent-openai
llama-index-cli
llama-index-core
llama-index-embeddings-huggingface
llama-index-embeddings-openai
llama-index-embeddings-ollama
llama-index-indices-managed-llama-cloud
llama-index-legacy
llama-index-llms-huggingface
llama-index-llms-llama-cpp
llama-index-llms-ollama
llama-index-llms-openai
llama-index-llms-langchain
llama-index-multi-modal-llms-openai
llama-index-program-guidance
llama-index-program-lmformatenforcer
llama-index-program-openai
llama-index-question-gen-openai
llama-index-readers-file
llama-index-readers-llama-parse
llama-index-readers-web
llama-index-readers-json
llama-index-readers-s3
llama-index-readers-azstorage-blob
llama-index-readers-microsoft-sharepoint
llama-index-readers-snowflake
llama-parse
llama_cpp_python
llamaindex-py-client
pydantic
python-dotenv
watchdog
img2table
paddleocr
duckduckgo-search
streamlit
Flask
anthropic
rich
openai
botocore
numpy
tavily-python
streamlit-aggrid
